% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.k.folds.R
\name{create.k.folds}
\alias{create.k.folds}
\title{Create Folds}
\usage{
create.k.folds(y, k = 10)
}
\arguments{
\item{y}{a vector of outcomes.}

\item{k}{an integer for the number of folds.}
}
\value{
List with the following elements:
                  out = list containing indices of the folds.
}
\description{
Create Folds
}
\examples{
create.k.folds()
}
\author{
Matthijs Knigge
}
\keyword{cross-validation}
\keyword{k-fold}
